Created Basic Layout
Used a basic HTML boilerplate as a base and used it to create the web page it also has most of the form done. Populated the CSS Files with styles and tried to create a 2 column form one with the customer info, the other with the computer configuration. Ran into problems need to resolve them.